pyproject.toml
[tool.poetry]
name = "science"
version = "0.5.0"
description = ""
authors = ["Vsevolod Bazhan <vsevozhan@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.0.5"
django-crispy-forms = "^1.9.0"
psycopg2-binary = "^2.8.5"
django-bootstrap-modal-forms = "^1.5.0"
django-widget-tweaks = "^1.4.8"
django-tables2 = "^2.3.1"
sortedcollections = "^1.1.2"
django-filter = "^2.2.0"
django-login-required-middleware = "^0.5.0"
tablib = {extras = ["xls"], version = "^2.0.0"}
gunicorn = "^20.0.4"
python-decouple = "^3.4"
dj-database-url = "^0.5.0"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
black = "^19.10b0"
isort = "^4.3.21"
coverage = "^5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
